+++ CHANGELOG.md
@@ -23,6 +23,14 @@
 
 -->
 
+## Unreleased
+
+### Breaking changes
+
+- Remove `set_global_time_context` function (https://github.com/shadowylab/universal-time/pull/1)
+- Link errors if the time provider is missing (https://github.com/shadowylab/universal-time/pull/1)
+- Require `define_time_provider!` macro on `no_std`/`WASM-unknown` (https://github.com/shadowylab/universal-time/pull/1)
+
 ## v0.1.0 - 2026/02/23
 
 First release.